remove redundant argument from FileMetaDataGeneratorProtocol.metadataForFile
