+++ wire-ios/Wire-iOS/Sources/UserInterface/Settings/CellDescriptors/SettingsCellDescriptorFactory+Advanced.swift
@@ -18,6 +18,7 @@
 
 import UIKit
 import WireSyncEngine
+import WireCommonComponents
 
 extension SettingsCellDescriptorFactory {
 
@@ -50,7 +51,8 @@ extension SettingsCellDescriptorFactory {
                 let viewModel = SettingsDebugReportViewModel(
                     router: router,
                     shareFile: shareFile,
-                    fetchShareableConversations: fetchShareableConversations
+                    fetchShareableConversations: fetchShareableConversations,
+                    fileMetaDataGenerator: FileMetaDataGenerator()
                 )
                 let viewController = SettingsDebugReportViewController(viewModel: viewModel)
                 router.viewController = viewController
